backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.12.1
sentence-transformers==2.7.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
pillow==10.1.0